refactor: Change the path of 'GameManager' and 'TriggerManager'
- /Games -> /Managers
