+++ Includes/Rosetta/Managers/GameManager.hpp
