+++ Includes/Rosetta/Managers/TriggerManager.hpp
